requirements.txt
Flask==2.0.1
gunicorn==20.1.0
roboflow==1.1.9
google-cloud-storage==2.0.0
Werkzeug==2.0.1
requests==2.31.0
python-dotenv==1.0.0